bump go to 1.24 (#166)
#159 is complaining about Go version used to build golangci-lint doesn't
match target version. That PR updates the toolchain to use Go 1.24.1,
but I believe golangci-lint is being built with one of the versions
[here](https://github.com/actions/runner-images/blob/ubuntu24/20250427.1/images/ubuntu/Ubuntu2404-Readme.md#go)
(likely `1.23.8`, based on the error message) which are included with
the `ubuntu-latest` image we're
[using](https://github.com/launchdarkly/find-code-references-in-pull-request/blob/main/.github/workflows/main.yml#L10).
The `Set up Go` job pulls the Go version from our go.mod file, so I'm
hoping moving that step above the pre-commit step will resolve this for
us.

Note: Also bumping to Go 1.24 here