bump viper incrementally (#168)
#159 attempts to bump `spf13/viper` from `1.19.0` to `1.20.1`, but there
is a failing GHA because the `e2e-tests` job gets the diff of the PR
from the GitHub API, and the diff is so huge that it exceeds the max
number of lines that can be returned from the GET pull request endpoint
(ripping out a bunch of unneeded code). So this PR bumps the version
incrementally to `1.20.0-alpha.1` which accounts for roughly half the
changes, and passes the `e2e-tests` job. We can then immediately upgrade
from `1.20.0-alpha.1` to `1.20.1` and I believe that diff should be
small enough as well. This means we'd be on an alpha version for a brief
time, which is a bit dicey, but I think acceptable. In any event, I'm
not sure how else we're going to get around this issue, unless we want
to dig into what the test is doing and try to come up with an acceptable
substitute, but that feels substantially riskier, actually. Thoughts?